test: address CodeRabbit review on PR #175
- OkHttpUtilsTest: drop misleading pool-capacity comment.
- RetryingCallbackTest: close OkHttpClient in tearDown to release dispatcher pool.
- HttpFunctionEndpointSpecTest: replace .isNotNull() with concrete IllegalArgumentException + message check.
- DefaultHttpRequestReplyClientFactoryTest: add @BeforeEach reset to guarantee clean singleton baseline per test.
- MatchBinderTest + StatefulMatchFunctionTest: extract shared NoOpContext, drop fully-qualified assertion, prune verbose narrative comments.

Tests still 27/27 (httpfn) and 11/11 (sdk-embedded match) green.
