+++ statefun-flink/statefun-flink-core/src/test/java/org/apache/flink/statefun/flink/core/httpfn/OkHttpUtilsTest.java
@@ -11,13 +11,11 @@ import org.junit.jupiter.api.Test;
 class OkHttpUtilsTest {
 
   @Test
-  void newClientHasUnboundedDispatcherAndPositiveConnectionPool() {
+  void newClientHasUnboundedDispatcherAndIdleConnectionPool() {
     OkHttpClient client = OkHttpUtils.newClient();
     try {
-      // Dispatcher must be set high enough to never throttle StateFun's invocation fanout.
       assertThat(client.dispatcher().getMaxRequests()).isEqualTo(Integer.MAX_VALUE);
       assertThat(client.dispatcher().getMaxRequestsPerHost()).isEqualTo(Integer.MAX_VALUE);
-      // Pool capacity is 1024; pin it so a future shrink to a lower value gets noticed.
       assertThat(client.connectionPool().connectionCount()).isZero();
       assertThat(client.followRedirects()).isTrue();
       assertThat(client.followSslRedirects()).isTrue();
@@ -47,8 +45,6 @@ class OkHttpUtilsTest {
 
     OkHttpUtils.closeSilently(client);
 
-    // Pin: a second close on the same client must NOT throw — closeSilently swallows all
-    // throwables from already-closed dispatchers/pools by design.
     assertThatCode(() -> OkHttpUtils.closeSilently(client)).doesNotThrowAnyException();
   }
 }
