+++ statefun-flink/statefun-flink-core/src/test/java/org/apache/flink/statefun/flink/core/httpfn/RetryingCallbackTest.java
@@ -49,6 +49,7 @@ class RetryingCallbackTest {
   @AfterEach
   void tearDown() throws IOException {
     server.shutdown();
+    OkHttpUtils.closeSilently(client);
   }
 
   @Test
