+++ statefun-sdk-embedded/src/test/java/org/apache/flink/statefun/sdk/match/MatchBinderTest.java
@@ -5,22 +5,16 @@ package org.apache.flink.statefun.sdk.match;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import org.apache.flink.statefun.sdk.Address;
-import org.apache.flink.statefun.sdk.Context;
-import org.apache.flink.statefun.sdk.io.EgressIdentifier;
-import org.apache.flink.statefun.sdk.metrics.Metrics;
 import org.junit.jupiter.api.Test;
 
+/**
+ * Pins the matching contract of {@link MatchBinder}. Production code reaches the binder only via
+ * {@link StatefulMatchFunction#invoke}, so these tests guard against silent precedence changes
+ * during refactoring.
+ */
 class MatchBinderTest {
-
-  // The MatchBinder constructor is package-private, so the test class lives in the same package
-  // and uses `new MatchBinder()` directly. Production code never news up a binder — it goes
-  // through StatefulMatchFunction.invoke → matcher.invoke. Tests pin the matching contract so a
-  // future refactor can't silently change precedence.
 
   @Test
   void typeOnlyPredicateMatchesByExactClassAndDispatchesAction() {
@@ -28,7 +22,7 @@ class MatchBinderTest {
     List<String> dispatched = new ArrayList<>();
     binder.predicate(String.class, (ctx, s) -> dispatched.add("string:" + s));
 
-    binder.invoke(new RecordingContext(), "hello");
+    binder.invoke(new NoOpContext(), "hello");
 
     assertThat(dispatched).containsExactly("string:hello");
   }
@@ -51,8 +45,8 @@ class MatchBinderTest {
     binder.predicate(
         String.class, s -> s.startsWith("a"), (ctx, s) -> dispatched.add("conditional:" + s));
 
-    binder.invoke(new RecordingContext(), "apple"); // matches conditional
-    binder.invoke(new RecordingContext(), "banana"); // falls through to type-only
+    binder.invoke(new NoOpContext(), "apple");
+    binder.invoke(new NoOpContext(), "banana");
 
     assertThat(dispatched).containsExactly("conditional:apple", "type-only:banana");
   }
@@ -65,9 +59,8 @@ class MatchBinderTest {
     binder.predicate(Integer.class, i -> i > 10, (ctx, i) -> dispatched.add(2));
     binder.predicate(Integer.class, i -> i > 100, (ctx, i) -> dispatched.add(3));
 
-    binder.invoke(new RecordingContext(), 1000);
+    binder.invoke(new NoOpContext(), 1000);
 
-    // Even though all three predicates would match, the FIRST one wins.
     assertThat(dispatched).containsExactly(1);
   }
 
@@ -76,7 +69,7 @@ class MatchBinderTest {
     MatchBinder binder = new MatchBinder();
     binder.predicate(String.class, (ctx, s) -> {});
 
-    assertThatThrownBy(() -> binder.invoke(new RecordingContext(), 42))
+    assertThatThrownBy(() -> binder.invoke(new NoOpContext(), 42))
         .isInstanceOf(IllegalStateException.class)
         .hasMessageContaining("Don't know how to handle");
   }
@@ -87,8 +80,8 @@ class MatchBinderTest {
     List<Object> defaulted = new ArrayList<>();
     binder.otherwise((ctx, msg) -> defaulted.add(msg));
 
-    binder.invoke(new RecordingContext(), 42);
-    binder.invoke(new RecordingContext(), "anything");
+    binder.invoke(new NoOpContext(), 42);
+    binder.invoke(new NoOpContext(), "anything");
 
     assertThat(defaulted).containsExactly(42, "anything");
   }
@@ -113,52 +106,20 @@ class MatchBinderTest {
     assertThatThrownBy(() -> binder.otherwise(null)).isInstanceOf(NullPointerException.class);
   }
 
+  /**
+   * Dispatch uses {@code input.getClass()} for exact match: a registration for {@link Number}
+   * does NOT catch {@link Integer}. Pin so users don't get a silent miss when expecting
+   * inheritance dispatch.
+   */
   @Test
   void typeOnlyMatchIsBasedOnExactClassNotInheritance() {
-    // Pin: dispatch uses input.getClass() exact match — a registration for Number does NOT
-    // catch Integer. Matters because users may expect inheritance dispatch.
     MatchBinder binder = new MatchBinder();
     List<Object> hits = new ArrayList<>();
     binder.predicate(Number.class, (ctx, n) -> hits.add(n));
 
-    assertThatThrownBy(() -> binder.invoke(new RecordingContext(), 42))
+    assertThatThrownBy(() -> binder.invoke(new NoOpContext(), 42))
         .isInstanceOf(IllegalStateException.class);
     assertThat(hits).isEmpty();
   }
 
-  /** Minimal Context implementation that records nothing — invoke() doesn't use the context. */
-  private static final class RecordingContext implements Context {
-    @Override
-    public Address self() {
-      return null;
-    }
-
-    @Override
-    public Address caller() {
-      return null;
-    }
-
-    @Override
-    public void send(Address to, Object message) {}
-
-    @Override
-    public <T> void send(EgressIdentifier<T> egress, T message) {}
-
-    @Override
-    public void sendAfter(Duration delay, Address to, Object message) {}
-
-    @Override
-    public void sendAfter(Duration delay, Address to, Object message, String cancellationToken) {}
-
-    @Override
-    public void cancelDelayedMessage(String cancellationToken) {}
-
-    @Override
-    public <M, T> void registerAsyncOperation(M metadata, CompletableFuture<T> future) {}
-
-    @Override
-    public Metrics metrics() {
-      return null;
-    }
-  }
 }
